Made compactTextBox much improved and more intuitive to implement. HUD is also better
Previously it would work only in certain cases, and utterly fail in others.
Now it doesn't.

The HUD now displays the hidden mana regen wait timer. It can also display "over-healing" up to 2x max health.

This is a new mechanic: A given healing spell or item has both the amount it heals,
  and the amount it can heal beyond your maximum health (over-healing).

 Over-healing effects should be primarily reserved for healing items.